doc autocomplete dropdown and move it closer to wehre its used
